Quality assurance fixes and PMD, error-prone integration

+++ modules/library/referencing/src/main/java/org/geotools/measure/SexagesimalConverter.java
@@ -193,7 +193,7 @@ class SexagesimalConverter extends AbstractConverter {
 
     @Override
     public Number convert(Number value) {
-        return convert(value);
+        return convert(value.doubleValue());
     }
 
     @Override
